requirements.txt
pytest
requests
allure-pytest
python-dotenvcl